clicking on non-walkable tiles finds the closest walkable tile and uses that as the dest added multi-requests (equip multiple things in the same tick for example) teleport explosions

+++ src/main/requests/AddExpRequest.java
@@ -9,12 +9,12 @@ public class AddExpRequest extends Request {
 		
 	}
 	
-	public AddExpRequest(int playerId, int statId, int exp) {
+	public AddExpRequest(int playerId, int statId, double exp) {
 		this.action = "addexp";
 		this.id = playerId;
 		this.statId = statId;
 		this.exp = exp;
 	}
 	private int statId;
-	private int exp;
+	private double exp;
 }
